+++ elements/pubsub/src/protocols/aps/types.ts
@@ -1,0 +1,8 @@
+import { Message } from '../types';
+
+export interface AccessDeniedMessage {
+  type: 'CHANNEL_ACCESS_DENIED';
+  payload?: any;
+}
+
+export type MessageData = Message | AccessDeniedMessage;